remade getUser-v2
