+++ src/pages/login/login.tsx
@@ -1,6 +1,6 @@
 import { FC, useState } from 'react';
 import { LoginUI } from '@ui-pages';
-import { loginUserThunk } from '../../services/slices/userSlice';
+import { loginUserThunk } from '../../services/actions';
 import { useAuthUser } from '../../hooks/useAuthUser';
 import { useDispatch, useSelector } from '../../services/store';
 import { getCookie } from '../../utils/cookie';
