+++ src/pages/profile-orders/profile-orders.tsx
@@ -2,10 +2,8 @@ import { ProfileOrdersUI } from '@ui-pages';
 import { TOrder } from '@utils-types';
 import { FC, useEffect } from 'react';
 import { useDispatch, useSelector } from '../../services/store';
-import {
-  getOrdersThunk,
-  selectOrdersList
-} from '../../services/slices/ordersListSlice';
+import { selectOrdersList } from '../../services/slices/ordersListSlice';
+import { getOrdersThunk } from '../../services/actions';
 
 export const ProfileOrders: FC = () => {
   const dispatch = useDispatch();
